+++ src/lib/constants/empty-room.ts
