+++ src/lib/constants/package-info.ts
