+++ src/lib/layout/components/auth/auth-popover.tsx
@@ -24,8 +24,8 @@ import { SpokerInput } from '~/lib/components/spoker-input';
 import { PRIVATE_ROUTES } from '~/lib/constants/routes/private';
 import { EVENT_TYPE_AUTH } from '~/lib/constants/tracking';
 import { logoutUser } from '~/lib/services/firebase/auth/logout';
-import { updateDisplayName } from '~/lib/services/firebase/auth/updateDisplayName';
-import { disconnectUser } from '~/lib/services/firebase/room/update/disconnectUser';
+import { updateDisplayName } from '~/lib/services/firebase/auth/update-display-name';
+import { disconnectUser } from '~/lib/services/firebase/room/update/disconnect-user';
 import { useAuth } from '~/lib/stores/auth';
 import { removeFirebasePrefix } from '~/lib/utils/removeFirebasePrefix';
 import { trackEvent } from '~/lib/utils/trackEvent';
