+++ src/lib/pages/room/components/task-list/index.tsx
@@ -30,13 +30,13 @@ import { ReactSortable } from 'react-sortablejs';
 import { AutoResizeTextarea } from '~/lib/components/auto-resize-textarea';
 import { SpokerModalWrapper } from '~/lib/components/spoker-modal-wrapper';
 import { SpokerWrapperGrid } from '~/lib/components/spoker-wrapper-grid';
-import { emptyRoom } from '~/lib/constants/emptyRoom';
-import { useUserRole } from '~/lib/hooks/useUserRole';
+import { emptyRoom } from '~/lib/constants/empty-room';
+import { useUserRole } from '~/lib/hooks/use-user-role';
 import { submitStoryFormValidationSchema } from '~/lib/models/task-list';
 import { editQueueItem } from '~/lib/services/firebase/room/update/queue/edit';
 import { removeQueueItem } from '~/lib/services/firebase/room/update/queue/remove';
 import { swapSelectedQueueWithCurrent } from '~/lib/services/firebase/room/update/queue/swap';
-import { rewriteQueue } from '~/lib/services/firebase/room/update/rewriteQueue';
+import { rewriteQueue } from '~/lib/services/firebase/room/update/rewrite-queue';
 import { useRoomStore } from '~/lib/stores/room';
 import type { Task } from '~/lib/types/raw-db';
