+++ src/lib/services/firebase/auth/get-current-user/index.ts
