+++ src/lib/services/firebase/auth/update-display-name/index.ts
