+++ src/lib/services/firebase/auth/verify-email/index.ts
