+++ src/lib/services/firebase/room/update/queue/edit/index.ts
@@ -1,4 +1,4 @@
-import { rewriteQueue } from '~/lib/services/firebase/room/update/rewriteQueue';
+import { rewriteQueue } from '~/lib/services/firebase/room/update/rewrite-queue';
 
 import type { EditQueueItemArgs } from './types';
