+++ src/lib/services/firebase/room/update/rewrite-completed.ts
