+++ src/lib/services/firebase/room/update/submit-vote/types.ts
