+++ tools/generate-rules.ts
